feat: Merge JavaScriptBootstrapUI into UI.
part of #15176
