+++ flow-server/src/main/java/com/vaadin/flow/component/UI.java
@@ -26,11 +26,24 @@ import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.regex.Pattern;
+import java.util.stream.Stream;
 
 import com.vaadin.flow.component.dependency.JsModule;
+import com.vaadin.flow.component.internal.AllowInert;
+import com.vaadin.flow.component.internal.JavaScriptNavigationStateRenderer;
+import com.vaadin.flow.component.page.History;
 import com.vaadin.flow.i18n.LocaleChangeEvent;
+import com.vaadin.flow.internal.nodefeature.NodeProperties;
+import com.vaadin.flow.router.ErrorNavigationEvent;
+import com.vaadin.flow.router.ErrorParameter;
+import com.vaadin.flow.router.NavigationEvent;
+import com.vaadin.flow.router.NavigationState;
+import com.vaadin.flow.router.NavigationStateBuilder;
 import com.vaadin.flow.router.NotFoundException;
+import com.vaadin.flow.router.RouteNotFoundError;
 import com.vaadin.flow.router.RouteParameters;
+import com.vaadin.flow.router.internal.ErrorStateRenderer;
+import com.vaadin.flow.router.internal.ErrorTargetEntry;
 import com.vaadin.flow.router.internal.HasUrlParameterFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -62,6 +75,7 @@ import com.vaadin.flow.router.QueryParameters;
 import com.vaadin.flow.router.RouteConfiguration;
 import com.vaadin.flow.router.Router;
 import com.vaadin.flow.router.RouterLayout;
+import com.vaadin.flow.router.internal.PathUtil;
 import com.vaadin.flow.server.Command;
 import com.vaadin.flow.server.ErrorEvent;
 import com.vaadin.flow.server.ErrorHandlingCommand;
@@ -69,8 +83,11 @@ import com.vaadin.flow.server.VaadinRequest;
 import com.vaadin.flow.server.VaadinService;
 import com.vaadin.flow.server.VaadinServlet;
 import com.vaadin.flow.server.VaadinSession;
+import com.vaadin.flow.server.auth.AnonymousAllowed;
 import com.vaadin.flow.server.communication.PushConnection;
 import com.vaadin.flow.shared.Registration;
+
+import elemental.json.JsonValue;
 
 /**
  * The topmost component in any component hierarchy. There is one UI for every
@@ -1112,7 +1129,7 @@ public class UI extends Component
      * @see #navigate(String)
      * @see Router#navigate(UI, Location, NavigationTrigger)
      *
-     * @param location
+     * @param pathname
      *            the location to navigate to, not {@code null}
      * @param queryParameters
      *            query parameters that are used for navigation, not
@@ -1120,14 +1137,50 @@ public class UI extends Component
      * @throws NullPointerException
      *             if the location or queryParameters are null.
      */
-    public void navigate(String location, QueryParameters queryParameters) {
-        Objects.requireNonNull(location, "Location must not be null");
+    public void navigate(String pathname, QueryParameters queryParameters) {
+        Objects.requireNonNull(pathname, "Location must not be null");
         Objects.requireNonNull(queryParameters,
                 "Query parameters must not be null");
+        Location location = new Location(pathname, queryParameters);
+        if (Boolean.TRUE.equals(getSession().getAttribute(SERVER_ROUTING))) {
+            // server-side routing
+            renderViewForRoute(location, NavigationTrigger.UI_NAVIGATE);
+            return;
+        }
 
-        getInternals().getRouter().navigate(this,
-                new Location(location, queryParameters),
-                NavigationTrigger.UI_NAVIGATE);
+        // client-side routing
+
+        // There is an in-progress navigation or there are no changes,
+        // prevent looping
+        if (navigationInProgress
+                || (getInternals().hasLastHandledLocation() && sameLocation(
+                        getInternals().getLastHandledLocation(), location))) {
+            return;
+        }
+
+        navigationInProgress = true;
+        try {
+            Optional<NavigationState> navigationState = getInternals()
+                    .getRouter().resolveNavigationTarget(location);
+
+            if (navigationState.isPresent()) {
+                // Navigation can be done in server side without extra
+                // round-trip
+                handleNavigation(location, navigationState.get(),
+                        NavigationTrigger.UI_NAVIGATE);
+                if (getForwardToClientUrl() != null) {
+                    // Server is forwarding to a client route from a
+                    // BeforeEnter.
+                    navigateToClient(getForwardToClientUrl());
+                }
+            } else {
+                // Server cannot resolve navigation, let client-side to
+                // handle it.
+                navigateToClient(location.getPathWithQueryParameters());
+            }
+        } finally {
+            navigationInProgress = false;
+        }
     }
 
     /**
@@ -1492,5 +1545,297 @@ public class UI extends Component
         } else {
             add(component);
         }
+    }
+
+    @Override
+    public Stream<Component> getChildren() {
+        // server-side routing
+        if (wrapperElement == null) {
+            return super.getChildren();
+        }
+
+        // #9069 with client-side routing, since routing component is a virtual
+        // child, its children need to be included separately (there should only
+        // be one)
+        Stream.Builder<Component> childComponents = Stream.builder();
+        wrapperElement.getChildren().forEach(childElement -> ComponentUtil
+                .findComponents(childElement, childComponents::add));
+        super.getChildren().forEach(childComponents::add);
+        return childComponents.build();
+    }
+
+    public static final String SERVER_ROUTING = "clientRoutingMode";
+
+    static final String SERVER_CONNECTED = "this.serverConnected($0)";
+    public static final String CLIENT_NAVIGATE_TO = "window.dispatchEvent(new CustomEvent('vaadin-router-go', {detail: new URL($0, document.baseURI)}))";
+
+    public Element wrapperElement;
+    private NavigationState clientViewNavigationState;
+    private boolean navigationInProgress = false;
+
+    private String forwardToClientUrl = null;
+
+    /**
+     * Gets the new forward url.
+     *
+     * @return the new forward url
+     */
+    public String getForwardToClientUrl() {
+        return forwardToClientUrl;
+    }
+
+    /**
+     * Connect a client with the server side UI. This method is invoked each
+     * time client router navigates to a server route.
+     *
+     * @param clientElementTag
+     *            client side element tag
+     * @param clientElementId
+     *            client side element id
+     * @param flowRoute
+     *            flow route that should be attached to the client element
+     * @param appShellTitle
+     *            client side title of the application shell
+     * @param historyState
+     *            client side history state value
+     */
+    @ClientCallable
+    @AllowInert
+    public void connectClient(String clientElementTag, String clientElementId,
+            String flowRoute, String appShellTitle, JsonValue historyState) {
+
+        if (appShellTitle != null && !appShellTitle.isEmpty()) {
+            getInternals().setAppShellTitle(appShellTitle);
+        }
+
+        final String trimmedRoute = PathUtil.trimPath(flowRoute);
+        if (!trimmedRoute.equals(flowRoute)) {
+            // See InternalRedirectHandler invoked via Router.
+            getPage().getHistory().replaceState(null, trimmedRoute);
+        }
+        final Location location = new Location(trimmedRoute);
+
+        if (wrapperElement == null) {
+            // Create flow reference for the client outlet element
+            wrapperElement = new Element(clientElementTag);
+
+            // Connect server with client
+            getElement().getStateProvider().appendVirtualChild(
+                    getElement().getNode(), wrapperElement,
+                    NodeProperties.INJECT_BY_ID, clientElementId);
+
+            getPage().getHistory().setHistoryStateChangeHandler(
+                    event -> renderViewForRoute(event.getLocation(),
+                            NavigationTrigger.CLIENT_SIDE));
+
+            // Render the flow view that the user wants to navigate to.
+            renderViewForRoute(location, NavigationTrigger.CLIENT_SIDE);
+        } else {
+            History.HistoryStateChangeHandler handler = getPage().getHistory()
+                    .getHistoryStateChangeHandler();
+            handler.onHistoryStateChange(new History.HistoryStateChangeEvent(
+                    getPage().getHistory(), historyState, location,
+                    NavigationTrigger.CLIENT_SIDE));
+        }
+
+        // true if the target is client-view and the push mode is disable
+        if (getForwardToClientUrl() != null) {
+            navigateToClient(getForwardToClientUrl());
+            acknowledgeClient();
+        } else if (isPostponed()) {
+            cancelClient();
+        } else {
+            acknowledgeClient();
+        }
+
+        // If this call happens, there is a client-side routing, thus
+        // it's needed to remove the flag that might be set in
+        // IndexHtmlRequestHandler
+        getSession().setAttribute(SERVER_ROUTING, Boolean.FALSE);
+    }
+
+    /**
+     * Check that the view can be leave. This method is invoked when the client
+     * router tries to navigate to a client route while the current route is a
+     * server route.
+     *
+     * This is only called when client route navigates from a server to a client
+     * view.
+     *
+     * @param route
+     *            the route that is navigating to.
+     */
+    @ClientCallable
+    public void leaveNavigation(String route) {
+        navigateToPlaceholder(new Location(PathUtil.trimPath(route)));
+
+        // Inform the client whether the navigation should be postponed
+        if (isPostponed()) {
+            cancelClient();
+        } else {
+            acknowledgeClient();
+        }
+    }
+
+    public void navigateToClient(String clientRoute) {
+        getPage().executeJs(CLIENT_NAVIGATE_TO, clientRoute);
+    }
+
+    private void acknowledgeClient() {
+        serverConnected(false);
+    }
+
+    private void cancelClient() {
+        serverConnected(true);
+    }
+
+    private void serverConnected(boolean cancel) {
+        wrapperElement.executeJs(SERVER_CONNECTED, cancel);
+    }
+
+    private void navigateToPlaceholder(Location location) {
+        if (clientViewNavigationState == null) {
+            clientViewNavigationState = new NavigationStateBuilder(
+                    getInternals().getRouter())
+                    .withTarget(ClientViewPlaceholder.class).build();
+        }
+        // Passing the `clientViewLocation` to make sure that the navigation
+        // events contain the correct location that we are navigating to.
+        handleNavigation(location, clientViewNavigationState,
+                NavigationTrigger.CLIENT_SIDE);
+    }
+
+    private void renderViewForRoute(Location location,
+            NavigationTrigger trigger) {
+        if (!shouldHandleNavigation(location)) {
+            return;
+        }
+        getInternals().setLastHandledNavigation(location);
+        Optional<NavigationState> navigationState = getInternals().getRouter()
+                .resolveNavigationTarget(location);
+        if (navigationState.isPresent()) {
+            // There is a valid route in flow.
+            handleNavigation(location, navigationState.get(), trigger);
+        } else {
+            // When route does not exist, try to navigate to current route
+            // in order to check if current view can be left before showing
+            // the error page
+            navigateToPlaceholder(location);
+
+            if (!isPostponed()) {
+                // Route does not exist, and current view does not prevent
+                // navigation thus an error page is shown
+                handleErrorNavigation(location);
+            }
+
+        }
+    }
+
+    private boolean shouldHandleNavigation(Location location) {
+        return !getInternals().hasLastHandledLocation()
+                || !sameLocation(getInternals().getLastHandledLocation(),
+                        location);
+    }
+
+    private boolean sameLocation(Location oldLocation, Location newLocation) {
+        return PathUtil.trimPath(newLocation.getPathWithQueryParameters())
+                .equals(PathUtil
+                        .trimPath(oldLocation.getPathWithQueryParameters()));
+    }
+
+    private void handleNavigation(Location location,
+            NavigationState navigationState, NavigationTrigger trigger) {
+        try {
+            NavigationEvent navigationEvent = new NavigationEvent(
+                    getInternals().getRouter(), location, this, trigger);
+
+            JavaScriptNavigationStateRenderer clientNavigationStateRenderer = new JavaScriptNavigationStateRenderer(
+                    navigationState);
+
+            clientNavigationStateRenderer.handle(navigationEvent);
+
+            forwardToClientUrl = clientNavigationStateRenderer
+                    .getClientForwardRoute();
+
+            adjustPageTitle();
+
+        } catch (Exception exception) {
+            handleExceptionNavigation(location, exception);
+        } finally {
+            getInternals().clearLastHandledNavigation();
+        }
+    }
+
+    private boolean handleExceptionNavigation(Location location,
+            Exception exception) {
+        Optional<ErrorTargetEntry> maybeLookupResult = getInternals()
+                .getRouter().getErrorNavigationTarget(exception);
+        if (maybeLookupResult.isPresent()) {
+            ErrorTargetEntry lookupResult = maybeLookupResult.get();
+
+            ErrorParameter<?> errorParameter = new ErrorParameter<>(
+                    lookupResult.getHandledExceptionType(), exception,
+                    exception.getMessage());
+            ErrorStateRenderer errorStateRenderer = new ErrorStateRenderer(
+                    new NavigationStateBuilder(getInternals().getRouter())
+                            .withTarget(lookupResult.getNavigationTarget())
+                            .build());
+
+            ErrorNavigationEvent errorNavigationEvent = new ErrorNavigationEvent(
+                    getInternals().getRouter(), location, this,
+                    NavigationTrigger.CLIENT_SIDE, errorParameter);
+
+            errorStateRenderer.handle(errorNavigationEvent);
+        } else {
+            throw new RuntimeException(exception);
+        }
+        return isPostponed();
+    }
+
+    private boolean isPostponed() {
+        return getInternals().getContinueNavigationAction() != null;
+    }
+
+    private void adjustPageTitle() {
+        // new title is empty if the flow route does not have a title
+        String newTitle = getInternals().getTitle();
+        // app shell title is computed from the title tag in index.html
+        String appShellTitle = getInternals().getAppShellTitle();
+        // restore the app shell title when there is no one for the route
+        if ((newTitle == null || newTitle.isEmpty()) && appShellTitle != null
+                && !appShellTitle.isEmpty()) {
+            getInternals().cancelPendingTitleUpdate();
+            getInternals().setTitle(appShellTitle);
+        }
+    }
+
+    private void handleErrorNavigation(Location location) {
+        NavigationState errorNavigationState = getInternals().getRouter()
+                .resolveRouteNotFoundNavigationTarget()
+                .orElse(getDefaultNavigationError());
+        ErrorStateRenderer errorStateRenderer = new ErrorStateRenderer(
+                errorNavigationState);
+        NotFoundException notFoundException = new NotFoundException(
+                "Couldn't find route for '" + location.getPath() + "'");
+        ErrorParameter<NotFoundException> errorParameter = new ErrorParameter<>(
+                NotFoundException.class, notFoundException);
+        ErrorNavigationEvent errorNavigationEvent = new ErrorNavigationEvent(
+                getInternals().getRouter(), location, this,
+                NavigationTrigger.CLIENT_SIDE, errorParameter);
+        errorStateRenderer.handle(errorNavigationEvent);
+    }
+
+    private NavigationState getDefaultNavigationError() {
+        return new NavigationStateBuilder(getInternals().getRouter())
+                .withTarget(RouteNotFoundError.class).build();
+    }
+
+    /**
+     * Placeholder view when navigating from server-side views to client-side
+     * views.
+     */
+    @Tag(Tag.DIV)
+    @AnonymousAllowed
+    public static class ClientViewPlaceholder extends Component {
     }
 }
